win32: test the decoder contract and the prediction width path
Follow-ups from the emoji fix review, folded into one change:

- test_terminal_width.cc: unit-test mosh_mbrtoc32's malformed-input
  contract — overlong 2/3/4-byte forms, lone continuation bytes,
  UTF-8-encoded surrogates, values above U+10FFFF, and invalid lead
  bytes all return (size_t)-1 with EILSEQ; a valid truncated prefix
  returns (size_t)-2; NUL consumes 0 with scalar 0; U+1F600 decodes to
  4 bytes and scalar 0x1F600.
- test_prediction_width.cc (new, wired into the native CI step): feed
  the four UTF-8 bytes of U+1F600 through PredictionEngine::new_user_byte
  and assert no one-cell prediction is installed (the pre-fix truncation
  to 0xF600 would have predicted a narrow PUA cell), with an ASCII 'a'
  control proving ordinary prediction still works. The Experimental
  display preference is load-bearing and documented: under Adaptive the
  epoch bump would make the control cell tentative and skip it.
- wincompat.{h,cc}: drop the mosh_char_t typedef; the shims take
  char32_t directly, and mosh_win32_wcwidth_scalar now takes char32_t
  too, finishing the canonical-scalar-type cleanup (the uint32_t casts
  in terminal.cc and terminaloverlay.cc go away).
- Makefile.win: clean removes the new test artifacts; the check comment
  covers both native-CI-only tests.
- PARITY.md: the astral Verification paragraph cites the
  decoder-contract asserts and test_prediction_width.

Sponsored-By: Netflix

+++ src/terminal/terminal.cc
@@ -45,7 +45,7 @@
    Width must be classified on that scalar, not on a 16-bit code unit. */
 static int wcwidth_scalar( char32_t ch )
 {
-  return mosh_win32_wcwidth_scalar( static_cast<uint32_t>( ch ) );
+  return mosh_win32_wcwidth_scalar( ch );
 }
 #endif
